requirements.txt
torch
torch-geometric
rdkit
scikit-learn
mlflow
pyyaml
numpy